pyproject.toml
[project]
name = "agents-course-gaia-challenge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "google-genai>=1.12.1",
    "httpx>=0.28.1",
    "llama-index-core>=0.12.33.post1",
    "llama-index-workflow>=1.0.1",
    "lxml>=5.4.0",
    "mediawikiapi>=1.3",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pydantic>=2.11.3",
    "pydantic-settings>=2.9.1",
    "tabulate>=0.9.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "mypy>=1.15.0",
]
